+++ packages/elec-ipc/CHANGELOG.md
@@ -9,10 +9,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.0.14](https://github.com/savage181855/savage-libs/compare/savage-electron-ipc@0.0.13...savage-electron-ipc@0.0.14) (2023-09-07)
+## [0.0.14](https://github.com/savage181855/savage-libs/compare/elec-ipc@0.0.13...elec-ipc@0.0.14) (2023-09-07)
 
-**Note:** Version bump only for package savage-electron-ipc
+**Note:** Version bump only for package elec-ipc
 
 ## 0.0.13 (2023-09-07)
 
-**Note:** Version bump only for package savage-electron-ipc
+**Note:** Version bump only for package elec-ipc