+++ packages/savage-types/CHANGELOG.md
@@ -9,10 +9,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.0.12](https://github.com/savage181855/data-types/compare/savage-data-types@1.0.11...savage-data-types@1.0.12) (2023-09-07)
+## [1.0.12](https://github.com/savage181855/data-types/compare/savage-types@1.0.11...savage-types@1.0.12) (2023-09-07)
 
-**Note:** Version bump only for package savage-data-types
+**Note:** Version bump only for package savage-types
 
 ## 1.0.11 (2023-09-07)
 
-**Note:** Version bump only for package savage-data-types
+**Note:** Version bump only for package savage-types